@@ -1,6 +1,6 @@
 {
   "name": "react-bootstrap-daterangepicker",
-  "version": "3.2.1-focus",
+  "version": "3.4.1",
   "description": "A date/time picker for react (using bootstrap). This is a react port of: https://github.com/dangrossman/bootstrap-daterangepicker",
   "main": "./lib/index.js",
   "scripts": {
@@ -18,36 +18,41 @@
   },
   "license": "Apache-2.0",
   "dependencies": {
-    "bootstrap-daterangepicker": "^2.1.24",
-    "object-assign": "^4.1.0"
+    "bootstrap-daterangepicker": "^2.1.25",
+    "object-assign": "^4.1.1"
   },
   "peerDependencies": {
+    "create-react-class": ">=15.5.0",
     "jquery": ">=2.0.0",
-    "react": ">=15.0.0"
+    "prop-types": ">=15.5.0",
+    "react": ">=15.5.0",
+    "react-dom": ">=15.5.0"
   },
   "devDependencies": {
-    "babel-preset-es2015": "^6.14.0",
-    "babel-preset-react": "^6.11.1",
-    "babel-preset-stage-0": "^6.5.0",
+    "babel-preset-es2015": "^6.24.1",
+    "babel-preset-react": "^6.24.1",
+    "babel-preset-stage-0": "^6.24.1",
     "babelify": "^7.3.0",
     "bootstrap": "^3.3.7",
     "cheerio": "^0.22.0",
-    "gulp": "^3.9.1",
+    "create-react-class": "^15.5.3",
     "gulp-babel": "^6.1.2",
     "gulp-browserify": "^0.5.1",
     "gulp-connect": "^5.0.0",
-    "gulp-cssmin": "^0.1.7",
+    "gulp-cssmin": "^0.2.0",
     "gulp-download": "0.0.1",
-    "gulp-less": "^3.1.0",
-    "gulp-nodemon": "^2.1.0",
+    "gulp-less": "^3.3.0",
+    "gulp-nodemon": "^2.2.1",
     "gulp-rename": "^1.2.2",
-    "gulp-uglify": "^2.0.0",
+    "gulp-uglify": "^2.1.2",
+    "gulp": "^3.9.1",
     "jsxhint": "^0.15.1",
-    "moment": "^2.14.1",
-    "react": "^15.3.1",
-    "react-bootstrap": "^0.30.3",
-    "react-dom": "^15.3.1",
-    "request": "^2.74.0",
+    "moment": "^2.18.1",
+    "prop-types": "^15.5.10",
+    "react-bootstrap": "^0.31.0",
+    "react-dom": "^15.5.4",
+    "react": "^15.5.4",
+    "request": "^2.81.0",
     "wordwrap": "^1.0.0"
   },
   "keywords": [
