@@ -6,15 +6,19 @@
  * Original copyright in: ./lib/daterangepicker.js
  */
 var React = require('react');
+var ReactDOM = require('react-dom');
 var $ = require('jquery');
 var objectAssign = require('object-assign');
 var DateRangePicker = require('bootstrap-daterangepicker');
 var getOptions = require('./get-options.js');
+var PropTypes = require('prop-types');
+var createReactClass = require('create-react-class');
 
 /* this is our export React class */
-module.exports = React.createClass({
+module.exports = createReactClass({
 	$picker: null,
 	options: getOptions(),
+	displayName: 'DateRange',
 	makeEventHandler: function (eventType) {
 		return function (event, picker) {
 			if (typeof this.props.onEvent === 'function') {
@@ -76,12 +80,14 @@ module.exports = React.createClass({
 		this.removeDateRangePicker();
 	},
 	removeDateRangePicker: function() {
-		this.$picker.data('daterangepicker').remove();
+		if(this.$picker && this.$picker.data('daterangepicker')) {
+			this.$picker.data('daterangepicker').remove();
+		}
 	},
 	initializeDateRangePicker: function() {
 		var $this = this;
 		$ = (window.jQuery && window.jQuery.fn.daterangepicker)? window.jQuery : $;
-		$this.$picker = $(this.refs.picker);
+		$this.$picker = $(ReactDOM.findDOMNode(this.refs.picker));
 		// initialize
 		$this.$picker.daterangepicker(this.getOptionsFromProps());
 		// attach event listeners
@@ -91,55 +97,55 @@ module.exports = React.createClass({
 		});
 	},
 	propTypes: {
-		"<input>": React.PropTypes.any,
-		alwaysShowCalendars: React.PropTypes.bool,
-		applyClass: React.PropTypes.string,
-		autoApply: React.PropTypes.bool,
-		autoUpdateInput: React.PropTypes.bool,
-		buttonClasses: React.PropTypes.array,
-		cancelClass: React.PropTypes.string,
-		dateLimit: React.PropTypes.object,
-		drops: React.PropTypes.oneOf(['down', 'up']),
-		endDate: React.PropTypes.oneOfType([
-			React.PropTypes.object,
-			React.PropTypes.string,
+		"<input>": PropTypes.any,
+		alwaysShowCalendars: PropTypes.bool,
+		applyClass: PropTypes.string,
+		autoApply: PropTypes.bool,
+		autoUpdateInput: PropTypes.bool,
+		buttonClasses: PropTypes.array,
+		cancelClass: PropTypes.string,
+		dateLimit: PropTypes.object,
+		drops: PropTypes.oneOf(['down', 'up']),
+		endDate: PropTypes.oneOfType([
+			PropTypes.object,
+			PropTypes.string,
 		]),
-		isCustomDate: React.PropTypes.bool,
-		isInvalidDate: React.PropTypes.func,
-		linkedCalendars: React.PropTypes.bool,
-		locale: React.PropTypes.object,
-		maxDate: React.PropTypes.oneOfType([
-			React.PropTypes.object,
-			React.PropTypes.string,
+		isCustomDate: PropTypes.func,
+		isInvalidDate: PropTypes.func,
+		linkedCalendars: PropTypes.bool,
+		locale: PropTypes.object,
+		maxDate: PropTypes.oneOfType([
+			PropTypes.object,
+			PropTypes.string,
 		]),
-		minDate: React.PropTypes.oneOfType([
-			React.PropTypes.object,
-			React.PropTypes.string,
+		minDate: PropTypes.oneOfType([
+			PropTypes.object,
+			PropTypes.string,
 		]),
-		onApply: React.PropTypes.func,
-		onCancel: React.PropTypes.func,
-		onEvent: React.PropTypes.func,
-		onHide: React.PropTypes.func,
-		onHideCalendar: React.PropTypes.func,
-		onShow: React.PropTypes.func,
-		onShowCalendar: React.PropTypes.func,
-		opens: React.PropTypes.oneOf(['left', 'right', 'center']),
-		parentEl: React.PropTypes.any,
-		ranges: React.PropTypes.object,
-		showCustomRangeLabel: React.PropTypes.bool,
-		showDropdowns: React.PropTypes.bool,
-		showISOWeekNumbers: React.PropTypes.bool,
-		showWeekNumbers: React.PropTypes.bool,
-		singleDatePicker: React.PropTypes.bool,
-		startDate: React.PropTypes.oneOfType([
-			React.PropTypes.object,
-			React.PropTypes.string,
+		onApply: PropTypes.func,
+		onCancel: PropTypes.func,
+		onEvent: PropTypes.func,
+		onHide: PropTypes.func,
+		onHideCalendar: PropTypes.func,
+		onShow: PropTypes.func,
+		onShowCalendar: PropTypes.func,
+		opens: PropTypes.oneOf(['left', 'right', 'center']),
+		parentEl: PropTypes.any,
+		ranges: PropTypes.object,
+		showCustomRangeLabel: PropTypes.bool,
+		showDropdowns: PropTypes.bool,
+		showISOWeekNumbers: PropTypes.bool,
+		showWeekNumbers: PropTypes.bool,
+		singleDatePicker: PropTypes.bool,
+		startDate: PropTypes.oneOfType([
+			PropTypes.object,
+			PropTypes.string,
 		]),
-		template: React.PropTypes.any,
-		timePicker: React.PropTypes.bool,
-		timePickerIncrement: React.PropTypes.number,
-		timePicker24Hour: React.PropTypes.bool,
-		timePickerSeconds: React.PropTypes.bool,
+		template: PropTypes.any,
+		timePicker: PropTypes.bool,
+		timePickerIncrement: PropTypes.number,
+		timePicker24Hour: PropTypes.bool,
+		timePickerSeconds: PropTypes.bool,
 	},
 	render: function () {
         if (this.$picker) {
@@ -152,6 +158,6 @@ module.exports = React.createClass({
 			delete props[key];
 		});
 
-		return React.createElement('div', props, this.props.children);
+		return React.cloneElement(this.props.children, {ref: 'picker'});
 	}
 });
